Add Sv2 binary codec that do not use serde
Add a new binary codec for the Sv2 format (no-serde-sv2). The new codec do not use
serde so it can be used in the bitcoin core codebase. The new binary
codec export an interface very similar to the serde-v2 one, so it can
be used almost in the same way, it actually export a sub-set of serde-sv2
interface/features so what is built with no-serde-sv2 shoud work with serde-sv2.

Add binary-sv2 that import both serde-sv2 and no-serde-sv2 and export a
shared interface via a compilation flag it underliaing codec can be
chose.

Upgrade all the crate that import serde-sv2 to work with the new
binary-sv2.