Add request id replacment logic, Fix sv2-ffi profile warning
Proxies need to change the downstream request_id in order to ensure that
request_id is connection-wide unique. Proxies also need a method to
reconcile the upstream request_id with the original downstream
request_id in oreder the route the upstream responses to the right
downstream.
This commit add the logic to do the above and also updated the proxy
example in order to explicitly show the request_id change.

sv2-ffi have a custom profile release cause it need to be called by C++
code. That custom profile release broke rust analyzer and rustc
version 1.57, it also genrate a warning. The costum profile is removed
and it must be replaced by the correct compiler flag in the
bitcoin-core build code.