Fix non copy datatype deserialization
The non serde parser when is deserializing data for non copy data type
(variable len data type) it check that the len of the passed vector is
less or equal than the maximum size for that datatype. That do not
consider the lenght of the fixed header (the one that encode the lenght)
so if a value of max size is passed the decoder fail.
This commit fix it and add several test for this particular case.