Decrease the github check for min codecoverage
This value is decreased as a lot of crates in this workspace are in
a very early stage or still to be developed. When the entire codebase
will be more mature this value must be increased.
That is justified by the fact that newly added code is "experimental"
and not worth to test as very likely to change, when the codebase will
be sufficently mature is really important to have a good code coverage.
Another solution would be to test code coverage only for the more stable
crates and ignore code coverage for the newly created ones.