Add logic to manage routing
Proxies, in order to route upstream's messages to the right downstream
connection, need to know which downtream connections belongs to a
specific group channel. This commit implement the logic needed to
accomplish that and update the proxy example to reflect the ability
for a proxy to chose the right downstream.

This commit make possible a complete round:
1. mining device -> proxy -> pool
2. pool -> proxy -> mining device