Add a safe Mutex
Add both a sync and an async Mutex that implements safe_lock and lock
where lock is the canonical Mutex API and safe_lock is the so called
closure API. The closure API take a closure as argument. The closure will
handle the value protected by the Mutex. Safe_lock return whatever the
closure return.
The closure API is superior to the canonical API cause it enfoce the drop
of the MutexGuard lowering the risk of deadlock.
The closure API is inferior to the canonical API cause can not be used
everywhere.
This commit also modify the code to use the closure API where it is
possible.