Refactor proxy logic
It moves the routing proxy logic in messages-sv2, cause it is not an
implementation detail and remove complexity from the proxy. Now the
various (sub)protocol message parser that are in messages::handlers
accept an enum that rapresent the routing logic so each implementation
will specify wich routing startegy apply chosing from the ones
avaiabeles in messages-sv2 (for now just None and Proxy).
Routing logic is removed from the struct definition of the nodes to
further reduce complexity.