Upgrade routing logic request_id mapper
When a proxy receive a message with a request_id field from downstream
it must ensure that the request_id of the relayed message to upstream is
connection-wide unique. In order to do that must updated the id of the
message and keep track of the original-id and the updated-id so that it
can transform the id of the response from upstream (that will be the
updated-id) into the original-id. Proxies when receive messages from
downstream do not clone anything, something of type Message is just a
view on the raw sv2 header + payload bytes. Beeing request-id an u32 so
a Copy type, changing it do not affect the original raw data.
This commit add a new type (a non copy u32) and use it for request-id(s)
it also add a method to updated the request-id(s) and change the old
request-id logic with a new one.