Update Extranonce.next
Before the next method had the following issue. Suppose that the extranonce is the following:
tail: MAX,
head:0.
If next method is called for the extranonce above, the tail is not resetted to 0:
tail: MAX,
head: 1.
So, the extranonce possibilities are MAX*2^128=2^129 insted of 2^256.

Now the next method uses all the 2^256 possibilities for the extranonce. This modification comes with its test