Update binary-sv2/serde-sv2
Update binary-sv2/serde-sv2 in order to work with the nowdays codebase.
We need it cause in the message generator we want to deserialize
messages from json and this can not be done with binary-sv2/no-serde-sv2.
It also update all the SRI codebase to work with the
binary-sv2/serde-sv2.
message-generator is excluded from the workspace cause if in the
workspace when we build another crate from the workspace in order to run
a test the crate get same dependecy of message-generator so it will end
up using the binary-sv2/no-serde-sv2 that is not what we want.