Improved MG test and documentation
Fixed some issue of some tests. Now the main pool-sri-test-1.json doesn
not call the mock of the template provider. This is beacuse the mock may
fail and the main test does not show it. Now the mocks need be started
separatedly. This need to be investigated.
Some SetupConnection messages are moved to a separated file.
Added actiondoc whenever needed and documentation.