Overall improvement of MG and tests (#419)
* Improved MG test and documentation

Fixed some issue of some tests. Now the main pool-sri-test-1.json doesn
not call the mock of the template provider. This is beacuse the mock may
fail and the main test does not show it. Now the mocks need be started
separatedly. This need to be investigated.
Some SetupConnection messages are moved to a separated file.
Added actiondoc whenever needed and documentation.

* Actiondoc, modularity

In a test it is no longer mandatory to include the actiondoc.
All the messages relative to SetupConnection are moved in the separate
module inside /test/message-generator/common-messages

* pool-sri-test and TP mock are improved

    What this tests does:
    1. Connects to the pool as a client
    2. Open a standard channel
    4. Receives a NewPrevHash and NewMiningJob from the pool
    3. Submit a share that actually is a solution of a fake bitcoin
       target
    4. Checks that a second NewPrevHash is received from the pool

    What TP mocks does:
    1. Connects to the pool as a server
    2. Waits SetCoinbaseOutputDatasize from the pool
    3. Sends NewTemplate and SetNewPrevHash to the pool
    4. Waits SubmitSolution from the pool and replies with a .Success
    5. Sends a NewPrevHash

    To make all of this work you must:
    a. Run the TP mock
    b. Run the pool with configuration file
/test/config/pool-mock-tp.toml
    c. Run the pool-sri-test-1.json

* Added JN mock and test with extended channel

Added the mock of Job Negotiator. Added a version of
pool-sri-test-1.json for exended channel. This new test is very easy and
does not have mocked roles.

* Reviewed

---------

Co-authored-by: darricksee <99291847+darricksee@users.noreply.github.com>