Fix option template_id for ShareMeetBitcoinTarget variant (channel_factory)
OnNewShare is an enum created when an upstream receive a new share. ShareMeetBitcoinTarget is the variant
created when the share is above or equal to the bitcoin target.
template_id is the id of the template used to create the job to which
the received share belongs. When a pool negotiate a job with downstream
(Job Negotiaor) the do pool do not have the template used to create the
job to which the share belong for that template_id must be an Option.