Update .codecov.yaml
This script will attempt to upload the coverage data using the Codecov Bash uploader, and if it fails (indicated by a non-zero exit status), it will retry up to 5 times with a 10 second delay between each attempt. If the upload is successful at any point during the retry loop, the loop will be terminated early.