sv1 benchmark (draft)
Remove distinction between v1 and v2 roles

Move sv1-mining-device in roles/test-utils

Add JDC job declarator client

Fix Cargo.lock

Add github actions also on branch dev

Update dev readme

Fix JDC must updated coinbase output on new token, Fix minor things

1. Pool output is communicated with the token so when JDC create a job
   it must use the the output of the token that is using
2. IS_NEW_TEMPLATE_HANDLED must be set to true after that JDC handle a
   new template received from the TP
3. JDC is a trasparent proxy. For that it Must use a PoolChannelFactory
   with:
   extranonce range 0 = 0..extranonce prefix len
   extranonce range 1 = range0.end..range0.end
   extranonce range 2 = range0.end..total extranonce len

   The above means the range1 len is 0 so the proxy do not own a search
   space and can not open multiple downstream channel, can only relay
   channel opened by the pool.

Add Sv1 client SetDiff, Add reinitialized JDC, Fix share sending JDC

_Add Sv1 client SetDiff_
Add a method called handle_set_difficulty to Sv1::IsClient. The sv1
mining device in test utils now update the upastream target based on the
difficulty that upstream require in order to get the desired number of
shares.

_Add reinitialized JDC_
If any async task of the JDC encounter a non recoverable error the JDC
abort all the async task and reinitialize itself. In the config file
there is a variable called `retry`, if JDC try to reinitialize itself for
a number of times bigger then `retry` the process exit.

_Fix share sending JDC_
If withould is on the JDC must send SubmitSolution to the TP and nothing to the
pool (when a share is below bitcoin target). Before this commit no message was
sent now we send SubmitSolution to the TP.

Add Sv1 client SetDiff, Add reinitialized JDC, Fix share sending JDC

_Add Sv1 client SetDiff_
Add a method called handle_set_difficulty to Sv1::IsClient. The sv1
mining device in test utils now update the upastream target based on the
difficulty that upstream require in order to get the desired number of
shares.

_Add reinitialized JDC_
If any async task of the JDC encounter a non recoverable error the JDC
abort all the async task and reinitialize itself. In the config file
there is a variable called `retry`, if JDC try to reinitialize itself for
a number of times bigger then `retry` the process exit.

_Fix share sending JDC_
If withould is on the JDC must send SubmitSolution to the TP and nothing to the
pool (when a share is below bitcoin target). Before this commit no message was
sent now we send SubmitSolution to the TP.

Add interop test for jdc

Remove JD capabilities from tproxy

Remove JD capabilities from mproxy

Add JDS, Remove JD capabilities from pool

rename JN/negotiation/negotiator to JD/declaration/declarator

into_static function for primitives

Implementation of into static for Seq064<'s, T>, with T is B016M, u32, ShortTxId.
For B032 it is necessary to distinguish between human readable or not
1) Rust internal representation --> binary: not human readable
2) Rust internal representation --> json: human readable
In order to change a sv2-message in the MG, it has to be converted in
serde_json::Value, modified, then in string and then deserialized again in a
sv2-message. Without this modification, the B032 is transformed in an
array in which the first element in the lenght of the vector that is in
the second element. This made impossible to deserialise the modified
message as string.

into_static function

a new file into_static.rs there is defined a new function called
into_static that transforms an Anymessage<'a> into a owned
Anymessage<'static>

Stateful test works

The test translation-proxy.json has a OpenExtendedMiningChannelSuccess
message with request_id = 666.i If you run it, you see that the message
is sent with the request_id field that is equal to the same field in the
message OpenExtendedMiningChannel previously received, namely 0.
This is done by a
a field "replace_fields" in the json test message. Replace fields is a
vector of objects of type ("field", "variable_name"). In this case, it is
just ("request_id", "variable_name"). In the actions, there is a new action
result called GetMessageField. This result is defined by a SV2 protocol, a
message_type and a vector of element of the form ("field", "variable_name").
Then it stores value = "field" and key = "variable_name" in an HashMap(key,
value). When a message non-empty replace field is sent, then the message
retrives from the HasMap the field values through the keys, that are the
second components of the elements of the vectors replace_fields.
So
GetMessageField --> HashMap(key, field_value) taken from a received message
Before sending a message, checks that for every ("field", "variable_name") there
there is some "field_value" for the key "variable_name" in the HasMap, so that in
the message that is going to be sent, the "field" is updated with "field_value".

Fix feature on the pool

In test pool-sri-test-reject-auth.json the pool is launched with the
ferature MG-reject-auth. A little modification to the pool toml is added
to make this feature work and the test pass.