Add Sv1 client SetDiff, Add reinitialized JDC, Fix share sending JDC
_Add Sv1 client SetDiff_
Add a method called handle_set_difficulty to Sv1::IsClient. The sv1
mining device in test utils now update the upastream target based on the
difficulty that upstream require in order to get the desired number of
shares.

_Add reinitialized JDC_
If any async task of the JDC encounter a non recoverable error the JDC
abort all the async task and reinitialize itself. In the config file
there is a variable called `retry`, if JDC try to reinitialize itself for
a number of times bigger then `retry` the process exit.

_Fix share sending JDC_
If withould is on the JDC must send SubmitSolution to the TP and nothing to the
pool (when a share is below bitcoin target). Before this commit no message was
sent now we send SubmitSolution to the TP.