into_static function for primitives
Implementation of into static for Seq064<'s, T>, with T is B016M, u32, ShortTxId.
For B032 it is necessary to distinguish between human readable or not
1) Rust internal representation --> binary: not human readable
2) Rust internal representation --> json: human readable
In order to change a sv2-message in the MG, it has to be converted in
serde_json::Value, modified, then in string and then deserialized again in a
sv2-message. Without this modification, the B032 is transformed in an
array in which the first element in the lenght of the vector that is in
the second element. This made impossible to deserialise the modified
message as string.