Stateful test works
The test translation-proxy.json has a OpenExtendedMiningChannelSuccess
message with request_id = 666.i If you run it, you see that the message
is sent with the request_id field that is equal to the same field in the
message OpenExtendedMiningChannel previously received, namely 0.
This is done by a
a field "replace_fields" in the json test message. Replace fields is a
vector of objects of type ("field", "variable_name"). In this case, it is
just ("request_id", "variable_name"). In the actions, there is a new action
result called GetMessageField. This result is defined by a SV2 protocol, a
message_type and a vector of element of the form ("field", "variable_name").
Then it stores value = "field" and key = "variable_name" in an HashMap(key,
value). When a message non-empty replace field is sent, then the message
retrives from the HasMap the field values through the keys, that are the
second components of the elements of the vectors replace_fields.
So
GetMessageField --> HashMap(key, field_value) taken from a received message
Before sending a message, checks that for every ("field", "variable_name") there
there is some "field_value" for the key "variable_name" in the HasMap, so that in
the message that is going to be sent, the "field" is updated with "field_value".