fix replace field bug
Before, if we tried to use ReplaceField in a message with a field that
is not a u32, it would generate an error because the function
change_value_of_serde_field was set to modify only fields that are u32.
Now it is fixed. You can see it by running the translation-proxy.json,
where several fields' types in messages are replaced.