Fix bugs on JD and MG
FIRST: in the implementation of GetSize of CommitMiningJob, there were
some bugs like "coninbase" in place of "coinbase". The compiling of this
part of software is disabled by default, unless compiling with the
feature "with_serde". This is why these issues are undiscovered.

SECOND: Some JD messages are added to the sri, so their correspectives
in the MG are added