Add key-utils
The new noise protocol used by SRI deprecate elliptic curve signing algorithm and replace it with
secp256k1, so that can be compatible with bitcoind. Key utils will replace noise_sv1::formats, and
can be used to serialize and deserialize secp256k1 keypairs in base58. It can also be used as an
utility to generate a random keypair.