Add some rust tests
The tests added all compile all except test_serialize_and_deserialize_6_dmj
this is no loss as there are already older tests that are not compiling,
probably inadvertently broken some PRs ago by someone.
since this pr is about arbitrary and it is already too long, i do not do
this now
We must include in the github actions something that performs all the
rust tests, btw. This is already stated in this issue
https://github.com/stratum-mining/stratum/issues/603