Add danger_set_start to the Buffer trait
New Sv2 framing consists of two or more encrypted messages. The first message
is the encrypted header, while the others are payload chunks, each split into
65,535-byte segments and encrypted. To build a buffer that can be deserialized
into an Sv2 message (header + payload), we need to call the decrypt function at
least twice on the same buffer: first for the header and then for the payload
chunks.

When passing a buffer to the decrypt function, it will attempt to decrypt the
entire buffer starting from the first byte. This behavior is correct only when
decrypting the header. In other cases, we need to "trick" the decrypt function
by making it see a buffer that starts from an index other than 0. For this, the
`danger_set_start` function is used.

Note: This function is currently unreliable and should be used with caution for
the following reasons:
1. It has not been fuzz-tested yet.
2. Users must remember to call `danger_set_start(0)` after decryption is
   complete.
3. While it works as expected with `as_ref` and `as_mut` methods, other
   specific methods still see the buffer starting at index 0.