JDsMempool::update_mempool and mempool in main
the method update_memppol of JDsMempool now is non-blocking. What we pay
is a bit of performance, as we have to generate a new rpc client every
time we update the mempool. To fix this requires a major refactoring of
the rpc client, that I cannot do for lack of time
the mempool is initialized in main and gets updated every second