improvements for the jd server
refactored the function that calculates the short_hash_id of a
tyransaction given the txid and a nonce. Refacored the method that takes
a mempool and a short tx id and verifies that there is that tx in the
mempool. Now gives Some(tx, txid) in positive case, none otherwise.

With all this machinery now we can identify the transactions sent in the
declared mining job. The identified transaction are stored in a field of
JobDeclaratorDownstream. It is also added a field that indicates how
many transactions are unidentified. In the case this field is non zero,
the JDS must send a ProvideMissingTransactions to the JDC.