fix division by zero and error management
the functions hash_rate_from_target and hash_rate_to_target performed
some divisions. If the denominator is zero this function panics and this
is bad. So these tu function return a Ok() for the happy path and an
Err() otherwise. Therefore, some work on error management it has been
added.