changed signatures of functions
If we use share per minute, we must perform divisions if we want to
obtain how many seconds to obtain a share. This expose to some risks as
the denominator can be zero. Changing the signature in this way shifts
the task of dividing to the caller, which is good.