Add logic for sender and receiver
Now the JDS, when receives the SubmitSolution, construct the block send
sends its hex serialization to the mempool. At the same time, an asyc
tast on main makes the mempool listen for new hex_encoded blocks. When
one arrives, it is submitted to the bitcoin node via the MiniRpcclient.
Note that this task of submitting a solution to the bitcoin node before
was blocking and now it is async.