easy script for compiling during development
after that the SRI has been divided in several workspaces for external
consuption, it is annoying to compile every workspace manually. So, a
script named clippy-on-all-workspaces.sh has been added. This script is
mean to aid the the final phase of developing a feature, where build
compiles and we pass to clippy.
In a preliminary phase of developing though, it is necessary something
that compiles a bit less deep and we do want to make tests pass. Thats
why I propose this script that does exactly this:
performes `cargo build` and `cargo +nightly fmt` in each workspace.