Add handlers for deserilized messages
Each Sv2 protocol have a server and client trait. This trait have a
method that handle serilaized messages. This commit add a method that
handle deserilized messages, and modify the method for serilaized so
that it just deserilize the message and call the new method. This can be
usefull if we have more roles managed by the same rust process, to avoid
some serailization and deserialization. It also bump the minor version
of the roles-logic-sv2 lib cause it add a feature and do not breack and
API.