fix message generator tests
Since the task that listen for new blocks is now performed
independently in main, if the url rpc config is the jds is left empty
the mempool the method `get_client` called in `JDsMemppol::on_submit()`
returns `NoClient` error. Since we do not want rpc connections in
message generator, the url config is left empty, resulting in a lot of
`NoCLient` errors in the JD-server, which led the MG tests involving
jd-server to fail.

Also changed some variables name to make them more intuitive