Fix network helper panic if peer close connection
When the remote peer try to open an encrypted sv2 connection and then
immediatly close the tcp connection, the helper for async std and tokio
it panic, mostly of the time we don't want to panic in this situation.
This commit add an error for this case and return it.