fix serialization of primitives with serde
some tests in main.rs of Message Generator are not passing or are
commented. This is because serialization and/or deserialization of Sv2
primitives was broken.

into_static for serde_sv2 primitives
these functions were panicking if the seq field of sequences (Seq<T> or
Sv2Option) were Some(...).

most rust tests for MG are fixed, one is added