Constrain `T` and `B` generics
`Frame` is defined with `T` and `B` generics but the constraints are
only introduced in the impl level which makes it harder to read the enum
and understand whats those generics are about. As those generics are a
key part of the `Frame` enum, it makes more sense to introduce the
constraints in the enum level.