impl: IdentifyTransactions.Success for jdc
We scrub the last declared mining jobs and collect a transaction list from the JD. Then mapping over the found transactions we convert the txid in the binary_sv2::U256 type. From this we can create a new Seq064K<U256> to send back from the JDC to the JDS that encountered a hash collision or was unable to creacreate the tx_hash_list_hash.