Add support for data types defined by sv2 extensions
In no-serde-sv2 an sv2 sequence is generic over T. If we want this
sequence to be Deserialize we need T to be Fixed and GerMarker. This 2
traits were private since all the sv2 types are already defined in
no-serde-sv2. But if we want to use sv2 types defined in an sv2
extensions we need to make these traits public.

The Encodable dervive macro in derive_codec implement GetSize for the
passed struct. But GetSize is also a blanket implementation for every
type that implement Fixed. So if we implement Fixed for our new sv2 type
and then we derive Encodable (commonly renamed Serialize) we get an
error. This commit add an attribute to Encodable called already_sized if the
struct that we want derive Encodable is market as already_sized the
macro will not implement GetSize for it.

This commit also bump minor version of derive_codec and no-serde-sv2
consequentially also of binary-sv2 since it reexport the above libs.