Merge `protocols` and `roles` workspaces
This commits removes the `protocols` and `roles` workspaces and creates
a new shared workspace for both of them. In practice we dont have a need
to maintain two different workspaces and no benefit is derived from
that.

With moving to a shared workspaces, the naming of the parent folders was
changed as will with `roles` becoming `stratum-roles` and `protocols`
becoming `stratum-protocols` and same for `common` crate. This required
a fix a across the codebase of the imports inside the different
`Cargo.toml`s.