Add `Sniffer` to `tests-integration` crate
Sniffer allows to intercept messages sent between two roles.

The sniffer sets between the downstream and upstream roles, aggregates
messages sent by the downstream and upstream roles in two separate FIFO
queues. The messages are then forwarded to the target role without any
modification.

It is useful for testing purposes, as it allows to assert that the roles
have sent specific messages in a specific order and to inspect the
messages details.