Update ping pong examples with handling of messages with additional data
Add exmaples on ping pong with and without noise, of how we can handle
messages that have additional field.
In the without noise example, the receiver is aware of the additional
field and can receive both messages with an wihout the additional field.
In order to do that it try **before** to parse the message with the
additional field, and if it fail it try to parse it without.

In the with noise example, we have a receiver that is only aware of the
message without the additional field, and a sender that send it with the
additional field, the sender ignore the additional field without
failing.