Handle Root Future Termination in Shutdown Scenarios
This is crucial because when the `start` method is executed
in a runtime that is not the main one, the lifecycle of futures
spawned within the `start` method continues, even after the `jdc`
process terminates.The earlier approach of simply exiting the
loop worked because the `start` method was executed on the main
blocking thread, where the Tokio runtime was defined. However,
in the case of integration tests, the runtime runs on a different
blocking thread, necessitating proper handling of the root
future termination.