Always use latest version of internal crates
In practice we always keep all of our crates updated with latest code
changes in all of other internal crates, so instead of maintaing the
versioning of each one for internal use, this commit just links to the
latest code in each crate.

This would alllow us to catch errors that might appear in one crate but
are originated in other. While this would sometimes require more work,
this can save us error that previously would appear only in the release
process.

This approach fits the current directory/project structre, but
if crates are moved outside of this repo, they should be linked to the
published crate on crates.io.