While the panic trace can be annoying..
..if the error originates from unhandled message(s), it is quite useful if
it is a different kind of error. Initially I thought removing the stack
trace would be cleaner and would result in better logs, but in practice
any error that occurs in testing is currently not printed, unless it is
an unhandled message.

This commit stops this behavior and will always print the stack trace in
any case of failure in testing.