Use stable cargo clippy in CI
Previously msrv was used which prevented us from using newer features
from later clippy versions although it does not affect our actual msrv.
Here we change msrv to stable to benifit from recent clippy features.

This change should be safe unless there is a regression in cargo clippy
i.e., a feature exist in 1.75 but removed in stable but this is very
unlikely.