Make `PoolSv2::start` unblocking
Currently the state loop at the end of `PoolSv2::start` is blocking,
it is wrapped with `tokio::task` so we can return from the `start`
function. This should give us more flexibelty in how we work with
`PoolSv2`. For example we can access the `PoolSv2` state directly
through the struct.