Move all non `Sniffer` functions outside..
..`sniffer.rs` file. `MessageAggregator` is moved to
`message_aggregator.rs` and other utility functions are moved to
`utils.rs`. This leaves the code inside `sniffer.rs` only with stuff
related to the `Sniffer` object.  The motivation behind this change is
to make the `sniffer.rs` more compact in order to add the next set of
features to it, specificly: Currently it acts only as man in the middle
between two roles. In next iteration it should also act as an downstream
or upstream role only. Which would allow us to use it in unit test more
easily.