protocols/v2/codec-sv2: fix a couple examples that weren't compiling
It seems that the `with_buffer_pool` feature is not additive, and causes
code to break when enabled, including examples that are bundled with
this project. To see this, run `cargo test --all-features --all-targets`
in the protocols/v2 directory.

This was introduced in https://github.com/stratum-mining/stratum/pull/170
in the early days of this project without review or comment.

This should be fixed but for now we just patch the examples to compile
so that I can move on with this PR.