protocols/v2/roles-logic-sv2: move coinbase_output into its own module
Over the next few commits I am going to move all the coinbase-related
stuff into this module, clean it up a bit, then move it wholesale over
to roles/roles-utils/config-helpers where I will combine it with the
other struct of the name CoinbaseOutput over there.

Currently the two structs exist because one is the "real" struct to be
used and one is the "serde" struct which is to be deserialized (I
think). However, this separation is currently badly incomplete: both
structures are the same and validation logic is duplicated.

It would be better if there were one structure and all the serde logic
were hidden in private modules related to that structure.