protocols/v2/roles-logic-sv2: use parse rather than from_str in coinbase_output
This is purely a stylistic thing and I'm happy to revert this commit,
but using .parse rather than from_str avoids needing the FromStr import
and also makes the formatting more consistent since you're always doing
a.x().y().z() rather than x(a).y().z().

On the other hand it adds more LOC and isn't consistent with the rest of
the codebase so I understand if you want to drop it.