move coinbase_output module from protocols to roles
There are some unit tests which cover the TryFrom impl between the two
CoinbaseOutput structs, which are now identical. Delete these. The
checks that are done at this conversion are also done at the point of
conversion to a ScriptBuf, so there is no risk from dropping them. (And
in the next PR the checks will be done when constructing a
CoinbaseOutput.)