roles/roles-utils/config-helpers: make invalid CoinbaseOutputs unconstructible
A few tests in roles/jd-server/src/lib/config.rs change from "cannot use
the config" to "cannot parse the config" which for the user will result
in a much faster failure. Unfortunately there isn't a nice way to match
on ConfigErrors that I found, so I asserted equality of the error
messages. I hope that's okay, even though they make the tests a bit
fragile.

I retained the CoinbaseOutput::new function for now; it is used in the
integration tests. But it is not used anywhere else and once we have a
descriptor-based constructor we will eliminate it.