roles/roles-utils/config-helpers: make invalid CoinbaseOutputs unconstructible
There is one test in roles/jd-server/src/lib/config.rs which checks
that, given a bad coinbase output, everything succeeds until the pool is
attempted to be started. I deleted this test because now you cannot
parse a configuration with a bad coinbase.

I renamed the CoinbaseOutput::new function to
CoinbaseOutput::from_type_and_value. This causes a bit of noise in this
diff. We are going to move away from this method of serializing coinbase
outputs and just keep it as a "legacy" encoding, so the longer name
makes sense. It's still useful for tests.