roles/roles-utils/config-helpers: add support for addr descriptors in coinbase output
This also adds an accessor `ok_for_mainnet` to `CoinbaseOutput` but it
does not use it anywhere. In a later PR we should hook this up and
refuse to start on mainnet if we are given a testnet address. It is
currently out of scope because it appears we have no way of determining
whether we are on mainnet or testnet (other than suggesting it in error
messages in case the user has an empty mempool).