roles/jd-server: refactor unit tests
In the next commit we will lose the CoinbaseOutput::new function, which
is used in several unit tests. We replace it with a new test helper
function which substitutes strings into an actual configuration TOML,
which is more flexible and is a better test.

This removes one redundant test but adds a couple new ones. The tests no
longer directly test the CoinbaseOutput::try_into method that yields a
ScriptBuf; we will remove that in the next commits in favor of "you
cannot construct a CoinbaseOutput that doesn't produce a ScriptBuf".