**Update protocol crates to support `State` cloning**
This commit introduces changes to the protocol crate to enable
the `State` to be clonable. This is necessary to allow the `State`,
once it transitions to the transport phase, to be used independently
on the reader and writer sides of the stream without requiring
synchronization primitives to share the state between them. This is
valid because, in the Noise protocol, the encryptor and decryptor
in the transport phase are mutually exclusive and do not need to
coexist within the same state instance. Although a more thorough
refactor could have been done to split the `noise_codec` into
separate `noise_codec_encryptor` and `noise_codec_decryptor`
components, we plan to revisit this as part of a future protocol
refactoring.